# server/openapi.yaml
openapi: 3.0.3
info:
  title: ProjectHelper Timeline API
  description: |
    프로젝트 타임라인(간트 차트) 데이터를 조작하는 REST API.
    AI 에이전트가 작업(Task) 단위로 일정을 읽고 수정할 수 있도록 설계됨.

    ## 동시성 (리비전)
    - 모든 변경 응답에 `revision`(정수)이 포함된다. 데이터가 바뀔 때마다 1씩 증가.
    - 변경 요청에 `If-Match: <revision>` 헤더를 보내면 서버 리비전과 다를 때 409를 반환한다.
    - 헤더를 생략하면 무조건 쓰기(last-write-wins) — 기존 클라이언트 하위호환.
    - 브라우저 앱은 10초마다 `GET /api/revision`을 폴링해 외부 변경을 자동 반영한다.

    ## 인증 (계정이 있을 때만)
    - 서버에 계정이 하나도 없으면 `open` 모드다 — 인증 없이 지금까지처럼 동작한다.
    - 첫 관리자를 만든 뒤에는 모든 경로에 신원이 필요하다. 예외는 정확히
      `/`, `/guide`, `/openapi.yaml`, `/health`, `/auth/*` 뿐이다.
    - 사람은 세션 쿠키(`POST /auth/login`), **에이전트는 서비스 토큰**을 쓴다:
      `Authorization: Bearer <token>` (서버의 `PH_API_TOKENS` 에 `이름:역할:토큰` 으로 등록).
      감사 로그의 actor 는 그 이름이 되므로 사람 계정을 빌려 쓰지 않는다.
    - 역할은 viewer(읽기) ⊂ editor(쓰기) ⊂ admin(계정·프로젝트 삭제).
      권한이 모자라면 403, 신원이 없으면 401.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 로컬 API 서버 (dev는 vite proxy 경유 :5173/api)
  - url: https://{host}/api
    description: Caddy HTTPS 프록시 (basicauth 필요)
    variables:
      host:
        default: localhost
security:
  - serviceToken: []
  - {}   # 계정이 없는 배포(open 모드)에서는 인증 없이 호출된다
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      description: PH_API_TOKENS 에 등록된 서비스 토큰. 사람 사용자는 세션 쿠키를 쓴다.
    session:
      type: apiKey
      in: cookie
      name: ph_session
  parameters:
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema: { type: string }
      description: 기대하는 현재 리비전. 불일치 시 409.
  schemas:
    Error:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error: { type: string }
        revision: { type: integer }
...
  responses:
    NotFound:
      description: 대상 없음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 리비전 불일치 (If-Match)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /tasks:
    get:
      summary: 작업 목록 (트리 또는 평탄화)
      parameters:
        - name: flat
          in: query
          schema: { type: string, enum: ['true'] }
          description: '"true"면 children 없이 level/parentId가 붙은 평탄 목록'
      responses:
        '200':
          description: '{ ok, revision, tasks }'
    post:
      summary: 작업 생성
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                parentId: { type: string, nullable: true }
                position: { type: integer, description: 형제 내 삽입 위치 (기본 맨 뒤) }
                startDate: { type: string, format: date, description: endDate와 함께 지정 시 첫 timeRange 생성 }
                endDate: { type: string, format: date }
                color: { type: string }
                description: { type: string }
                labels: { type: array, items: { type: string } }
      responses:
        '201': { description: '{ ok, revision, task }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: 작업 단건 조회
      responses:
        '200': { description: '{ ok, revision, task, parentId }' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: 작업 필드 부분 수정 (name/color/description/expanded/labels/divider/progress)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                color: { type: string }
                description: { type: string }
                expanded: { type: boolean }
                labels: { type: array, items: { type: string } }
                divider: { type: object }
                progress: { type: integer, minimum: 0, maximum: 100 }
      responses:
        '200': { description: '{ ok, revision, task }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: 작업 삭제 (하위 작업 포함)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '200': { description: '{ ok, revision }' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{id}/move:
    post:
      summary: 작업 이동 (재부모화/순서 변경)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parentId]
              properties:
                parentId: { type: string, nullable: true, description: null이면 루트 레벨 }
                position: { type: integer }
      responses:
        '200': { description: '{ ok, revision, task }' }
        '400': { description: 자기 서브트리로 이동 시 }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{id}/time-ranges:
    post:
      summary: 기간(바) 추가 — 작업의 startDate/endDate 캐시 자동 재계산
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startDate, endDate]
              properties:
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
                label: { type: string }
                color: { type: string, nullable: true }
                dependencies:
                  type: array
                  items: { type: string }
                  description: >
                    선행 엔티티 id (작업/기간/마일스톤). 존재하지 않는 id 이거나
                    순환을 닫는 연결이면 400. 점검은 GET /dependency-issues.
      responses:
        '201': { description: '{ ok, revision, timeRange }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{id}/time-ranges/{rangeId}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
      - name: rangeId
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: 기간 수정 (날짜/라벨/색/의존성)
      description: >
        cascade=true 면 이 기간의 종료일이 **뒤로 밀린 일수만큼** 후행(전이적 포함)
        전체를 같은 쓰기에서 민다 — 리비전 하나, 감사 한 줄. 앞당김(delta < 0)은
        전파하지 않는다: 사람이 잡아 둔 간격을 지우는 쪽이 더 큰 손실이다.
        응답에 cascaded(움직인 엔티티 id 목록)와 cascadeDays 가 붙는다.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - name: cascade
          in: query
          schema: { type: boolean, default: false }
        - name: workdays
          in: query
          schema: { type: boolean, default: false }
          description: cascade 로 옮긴 날짜가 주말이면 다음 평일로 (공휴일 달력은 없다)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                startDate: { type: string, format: date }
                endDate: { type: string, format: date }
                label: { type: string }
                color: { type: string, nullable: true }
                dependencies: { type: array, items: { type: string } }
      responses:
        '200': { description: '{ ok, revision, timeRange, cascaded?, cascadeDays? }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: 기간 삭제
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '200': { description: '{ ok, revision }' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{id}/milestones:
    post:
      summary: 마일스톤 추가
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date]
              properties:
                date: { type: string, format: date }
                label: { type: string }
                color: { type: string }
                shape: { type: string, enum: [diamond, circle, triangle, square, star, flag] }
                labelPosition: { type: string, enum: [auto, top, bottom, left, right] }
                dependencies: { type: array, items: { type: string } }
      responses:
        '201': { description: '{ ok, revision, milestone }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /tasks/{id}/milestones/{milestoneId}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
      - name: milestoneId
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: 마일스톤 수정 (날짜/라벨/도형/색/라벨위치/의존성)
      description: >
        **삭제 후 재생성으로 대신하지 마라** — id 가 바뀌고, 이 마일스톤을 가리키던
        dependencies 는 삭제 시점에 함께 정리되므로(pruneDependencies) 연결이 사라진다.
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                date: { type: string, format: date }
                label: { type: string }
                color: { type: string }
                shape: { type: string, enum: [diamond, circle, triangle, square, star, flag] }
                labelPosition: { type: string, enum: [auto, top, bottom, left, right] }
                dependencies: { type: array, items: { type: string } }
      responses:
        '200': { description: '{ ok, revision, milestone }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      summary: 마일스톤 삭제
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      responses:
        '200': { description: '{ ok, revision }' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /critical-path:
    get:
      summary: 임계경로와 여유(slack)
      description: >
        날짜는 이미 사람이 정한 값이므로 전진 계산(가장 이른 시작)을 하지 않는다 —
        후진 계산만으로 "이것이 며칠 늦어도 프로젝트 종료가 밀리지 않는가"를 답한다.
        LF(n) = min(후행들의 LS), 후행이 없으면 프로젝트 종료일.
        slack 0 이하가 임계다. 순환이 있으면 위상 순서가 없으므로 400 —
        /dependency-issues 로 먼저 고쳐야 한다. 주말은 여유로 세지 않은
        slackWorkdays 를 함께 준다(공휴일 달력은 없다).
      responses:
        '200':
          description: >
            { ok, revision, projectEnd,
            nodes: [{id, name, type, start, end, latestStart, latestFinish,
            slackDays, slackWorkdays, critical}],
            criticalIds: [id] }
        '400': { $ref: '#/components/responses/BadRequest' }
  /chart:
    get:
      summary: 텍스트 간트 렌더 — 에이전트가 자기가 쓴 결과를 확인하는 수단
      description: >
        AI 는 화면을 볼 수 없으므로, 쓰기만 있고 그림이 없으면 결과가 읽을 만한지
        판정할 수단이 0 이다. 대량 편집 뒤 한 번 호출해 막대 위치를 확인한다.
      parameters:
        - name: width
          in: query
          schema: { type: integer, minimum: 40, maximum: 400, default: 100 }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: maxRows
          in: query
          schema: { type: integer, minimum: 1, maximum: 1000, default: 200 }
        - name: format
          in: query
          schema: { type: string, enum: [json, text], default: json }
          description: text 면 text/plain 본문만 (JSON 이스케이프 없이 그대로 읽힌다)
      responses:
        '200': { description: '{ ok, revision, chart } — format=text 면 text/plain' }
  /batch:
    post:
      summary: 여러 변경을 한 번의 쓰기로 (전부 아니면 전무)
      description: >
        ops 를 순서대로 **같은 트랜잭션** 안에서 적용한다 — 리비전 하나, 감사 한 줄,
        하나가 실패하면 앞선 것까지 되돌린다. 단건 엔드포인트와 **같은 뮤테이터**를
        쓰므로 검증·순환 차단이 동일하다. blob POST /data 를 대체하는 경로다:
        저것은 트리 전체를 덮어써서 남의 편집을 지우지만, 이것은 조작만 보낸다.
        ref: 이 op 이 만든 것을 뒤에서 가리킬 이름. 뒤의 op 은 `@이름`(작업 id) /
        `@이름:range`(그 작업의 첫 기간 id)로 참조한다 — 부모-자식 트리와 연결을
        한 번에 만들 수 있다. 최대 200개.
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ops]
              properties:
                ops:
                  type: array
                  minItems: 1
                  maxItems: 200
                  items:
                    type: object
                    required: [op]
                    properties:
                      op:
                        type: string
                        enum: [create-task, update-task, delete-task, move-task,
                               add-time-range, update-time-range, delete-time-range,
                               add-milestone, update-milestone, delete-milestone]
                      ref: { type: string, pattern: '^[A-Za-z0-9_-]+$' }
                      taskId: { type: string }
                      rangeId: { type: string }
                      milestoneId: { type: string }
                      body: { type: object }
            example:
              ops:
                - op: create-task
                  ref: design
                  body: { name: 설계, startDate: "2026-01-01", endDate: "2026-01-31" }
                - op: create-task
                  body: { name: 요구사항 정리, parentId: "@design", startDate: "2026-01-01", endDate: "2026-01-10" }
                - op: create-task
                  ref: dev
                  body: { name: 개발, startDate: "2026-02-01", endDate: "2026-03-31" }
                - op: update-time-range
                  taskId: "@dev"
                  rangeId: "@dev:range"
                  body: { dependencies: ["@design:range"] }
      responses:
        '200': { description: '{ ok, revision, applied, results: [...] }' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...